hr-medbot-backend-master/hr-medbot-backend-master/application/requirements.txt
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pgvector==0.4.1
requests==2.32.3
alembic==1.15.2
openai==1.78.1
fastapi==0.115.12
uvicorn==0.34.2
bcrypt==4.3.0
pyjwt==2.10.1
python-multipart==0.0.20
langchain_text_splitters==0.3.8

websockets==15.0.1
httpx==0.28.1
diskcache==5.6.3